requirements.txt
fastapi
uvicorn[standard]
python-multipart
mediapipe
opencv-python-headless
numpy
python-dotenv
openai-whisper
praat-parselmouth
google-generativeai